be-identity/pyproject.toml
[tool.poetry]
name = "identity"
version = "0.1.0"
description = "Identity node"
authors = ["Jason Colburne <j.colburne@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.78.0"
uvicorn = "^0.18.2"
gunicorn = "^20.1.0"
toml = "^0.10.2"
asyncpg = "^0.26.0"
sqlmodel = "^0.0.6"
alembic = "^1.8.0"
python-multipart = "^0.0.5"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["argon2"], version = "^1.7.4"}
pydantic = {extras = ["email"], version = "^1.9.1"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[tool.poetry.scripts]
start = "common.main:start"
migrate = "database:migrate"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
